+++ CHANGELOG.rst
@@ -19,6 +19,8 @@ HEAD
   - add vcs history tree to log (prints the context of the vcs tree)
   - remodel the output of the degradation checks
   - switch the colour of optimizations to green (instead of blue)
+  - colour tainted (containing degradation) and fixed (containing optimization) branches in vcs history
+  - add short summary of degradations to each minor version in graph
 
 0.14.1 (2018-04-19)
 -------------------